pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-rag-chatbot"
version = "0.1.0"
description = "Chatbot legal con RAG (Leyes laborales argentinas)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Lucas", email = "tuemail@ejemplo.com"}]
license = {text = "MIT"}
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",   # ← importante
    "openai>=1.0.0",
    "pinecone-client>=3.0.0",
    "boto3>=1.34.0",
    "streamlit>=1.35.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[project.optional-dependencies]
dev = ["pre-commit", "mypy", "pytest-cov"]